Testing iteration weight being applied to simc

+++ src/main_iter.py
@@ -3,7 +3,7 @@
 #
 # Description:
 # ================================================================
-# Time-stamp: "2023-09-18 12:15:25 trottar"
+# Time-stamp: "2023-09-18 12:18:10 trottar"
 # ================================================================
 #
 # Author:  Richard L. Trotta III <trotta@cua.edu>
@@ -188,10 +188,12 @@ from compare_simc_iter import compare_simc
 # ***Moved up in procedure vs main.py since required for weight iteration
 new_dir = CACHEPATH+"/"+USER+"/"+ParticleType.lower()+"/"+formatted_date
 create_dir(new_dir)
+
 # ***Parameter file from last iteration!***
 # ***These old parameters are needed for this iteration. See README for more info on procedure!***
 old_param_file = '{}/src/{}/parameters/par.{}_{}.dat'.format(LTANAPATH, ParticleType, pol_str, Q2.replace("p",""))
-old_fort_param = '{}/param_{}_{}.f'.format(prev_iter_dir, ParticleType, pol_str)    
+old_fort_param = '{}/param_{}_{}.f'.format(prev_iter_dir, ParticleType, pol_str)
+
 # Upate hist dictionary with effective charge and simc histograms
 for hist in histlist:
     # SIMC file with weight from last iteration
@@ -199,13 +201,18 @@ for hist in histlist:
     new_simc_root = old_simc_root.replace(closest_date, formatted_date)
     # Make sure old simc root file exists
     if os.path.exists(old_simc_root):
-        # Copy to new iteration so and then edit the weight
-        print("Copying {} to {}".format(old_simc_root, new_simc_root))
-        shutil.copy(old_simc_root,new_simc_root)
-        # Function to calculation new weight and apply it to simc root file 
-        iter_weight(old_param_file, old_fort_param, new_simc_root, inpDict, hist["phi_setting"])
-        hist.update(compare_simc(new_simc_root, hist, inpDict))
-    
+        # Make sure new simc root file exists
+        if os.path.exists(new_simc_root):
+            # Copy to new iteration so and then edit the weight
+            print("Copying {} to {}".format(old_simc_root, new_simc_root))
+            shutil.copy(old_simc_root,new_simc_root)
+            # Function to calculation new weight and apply it to simc root file 
+            iter_weight(old_param_file, old_fort_param, new_simc_root, inpDict, hist["phi_setting"])
+            hist.update(compare_simc(new_simc_root, hist, inpDict))
+        else:
+            print("ERROR: {} not properly copied to {}".format(old_simc_root, new_simc_root))
+            sys.exit(2)
+            
 sys.path.append("plotting")
 from data_vs_simc import plot_data_vs_simc
